Add name selection implementation
